templates/starter/pyproject.toml
[project]
name = "my-ai-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "strands-agents",
    "bedrock-agentcore",
]
